Added python script for generating synthetic rain data (1 image at a time)
It appears that the c++ code can correctly estimate the distributions most of the time, but it is a bit fickle at times.

It is also appears that when it matches the distribution quite well it doesnt always pass the goodness-of-fit test. Maybe there is a bug there, or the threshold is maybe just very strict?