Started cleaning a bit up in the code to make it production ready.
Would it be usefull to have a verbose and debug tag, depending on what kind of information you want. Otherwise verbose will jsut clutter the cmd immediatly.

Added a frame counter, so that every 100th frame a upate is written, to keep track of how far the video procesing is. Will be written even if verbose is turned off.