Added a separate debug and verbose flag, and changed saveDebugImg to jsut saveImg. difference between verbose and debug is that verbose is if you want to follow the status and see EM values etc. while Debug also prints a lot more fro mthe or loops.
Made a separate function for plotting CDFS. It is a bit ugly solution but it works as of now, and results in a bit faster code as we dont have to compute the CDF plots all the time.

Updated header file description of algorithm.

Nearly ready production ready. Just needs to decide what we should print in the output file. Currently the code is a bit slower than I anticipated, evlauting 100 frame in ~1.5 minute in visual studio in release mode and all flags set to false, and with other process running on pc